move sepana utils to lib
