+++ src/lib/sepana/endpoints.ts
@@ -1,0 +1,6 @@
+export const SEPANA_ENDPOINTS = {
+  search: '/search',
+  delete: '/engine/data/delete',
+  insert: '/engine/insert_data',
+  job: (jobId: string) => `/job/status/${jobId}`,
+}
